requirement.txt
amqp==2.5.1
authy==2.2.3
autopep8==1.4.4
billiard==3.6.1.0
CacheControl==0.12.5
cachetools==3.1.1
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.1.5
django-cors-headers==3.0.1
django-debug-toolbar==2.0
django-debugging==0.1a2
django-filter==2.1.0
django-silk==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
firebase-admin==2.18.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gprof2dot==2017.9.19
grpcio==1.23.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
importlib-metadata==0.19
Jinja2==2.10.1
kombu==4.6.4
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pyfcm==1.4.7
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.1
requests==2.22.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
